Zeroday/requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1
matplotlib==3.7.2
seaborn==0.12.2
python-dateutil==2.8.2
pyyaml==6.0
requests==2.31.0
psutil==5.9.5
watchdog==3.0.0
colorama==0.4.6
tabulate==0.9.0

